homework while ill

+++ HOMEWORK_3/task_23/Program.cs
@@ -1,7 +1,7 @@
 ﻿Console.Write("Input power ");
 double pow = Convert.ToDouble(Console.ReadLine());
 Console.Write("Input amount ");
-double num = Convert.ToInt32(Console.ReadLine());
+int num = Convert.ToInt32(Console.ReadLine());
 RaiseToPower(num, pow);
 
 void RaiseToPower (double num, double pow)
